+++ apps/connect/src/components/app.tsx
@@ -5,10 +5,7 @@ import {
   DocumentEditorDebugTools,
   serviceWorkerManager,
 } from "@powerhousedao/connect/utils";
-import {
-  ToastContainer,
-  WagmiContext,
-} from "@powerhousedao/design-system/connect";
+import { ToastContainer } from "@powerhousedao/design-system/connect";
 import { useEffect } from "react";
 import { PackageInstallPrompt } from "./package-install-prompt.js";
 export const App = () => {
@@ -36,12 +33,10 @@ export const App = () => {
 
   return (
     <SentryProvider>
-      <WagmiContext>
-        <ToastContainer position="bottom-right" containerId="connect" />
-        <Router />
-        <PackageInstallPrompt />
-        <Analytics />
-      </WagmiContext>
+      <ToastContainer position="bottom-right" containerId="connect" />
+      <Router />
+      <PackageInstallPrompt />
+      <Analytics />
     </SentryProvider>
   );
 };