+++ packages/design-system/src/connect/components/modal/settings-modal/index.ts
@@ -1,5 +1,4 @@
 export * from "./clear-storage-row.js";
 export * from "./dependency-versions/dependency-versions.js";
-export * from "./document-select-row.js";
 export * from "./row.js";
 export * from "./settings-modal.js";