+++ packages/design-system/src/connect/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from "./use-drop.js";
 export * from "./use-drop.js";
+export * from "./use-ens.js";