+++ packages/reactor-browser/src/hooks/document-model-modules.ts
@@ -3,7 +3,17 @@ import { useVetraPackages } from "./vetra-packages.js";
 
 export function useDocumentModelModules(): DocumentModelModule[] | undefined {
   const vetraPackages = useVetraPackages();
-  return vetraPackages.flatMap((pkg) => pkg.documentModels);
+  return vetraPackages
+    .flatMap((pkg) => pkg.documentModels)
+    .filter(
+      (module, index, modules) =>
+        // deduplicate by documentType and version
+        modules.findIndex(
+          (m) =>
+            m.documentModel.global.id === module.documentModel.global.id &&
+            m.version === module.version,
+        ) === index,
+    );
 }
 
 export function useDocumentModelModuleById(